requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
openai==1.3.7
google-generativeai==0.3.2
elevenlabs==0.2.26
moviepy==1.0.3
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
pillow>=10.0.0
numpy>=1.24.0
aiohttp>=3.8.0 